feat(portfolio): qty 소수점 지원 (미국 fractional shares)

+++ src/portfolio.py
@@ -26,7 +26,7 @@ CREATE TABLE IF NOT EXISTS portfolio (
     username   TEXT NOT NULL,
     symbol     TEXT NOT NULL,
     avg_price  REAL NOT NULL,
-    qty        INTEGER NOT NULL DEFAULT 0,
+    qty        REAL NOT NULL DEFAULT 0,
     added_at   INTEGER NOT NULL,
     notes      TEXT,
     PRIMARY KEY (username, symbol)
@@ -132,15 +132,15 @@ def init_db() -> None:
     logger.info("portfolio DB 초기화 완료: %s", _DB_PATH)
 
 
-def _validate(avg_price: float, qty: int) -> None:
+def _validate(avg_price: float, qty: float) -> None:
     if avg_price is None or float(avg_price) <= 0:
         raise ValueError(f"avg_price must be positive, got {avg_price!r}")
-    if qty is None or int(qty) < 0:
+    if qty is None or float(qty) < 0:
         raise ValueError(f"qty must be non-negative, got {qty!r}")
 
 
 def add_holding(
-    username: str, symbol: str, avg_price: float, qty: int,
+    username: str, symbol: str, avg_price: float, qty: float,
     notes: str | None = None,
 ) -> bool:
     """추가 또는 갱신. 새로 추가됐으면 True, 기존 갱신이면 False."""
@@ -161,7 +161,7 @@ def add_holding(
                          avg_price = excluded.avg_price,
                          qty       = excluded.qty,
                          notes     = excluded.notes""",
-                    (username, symbol, float(avg_price), int(qty), now, notes),
+                    (username, symbol, float(avg_price), float(qty), now, notes),
                 )
                 conn.execute("COMMIT")
             except Exception:
@@ -190,7 +190,7 @@ def remove_holding(username: str, symbol: str) -> bool:
 def update_holding(
     username: str, symbol: str, *,
     avg_price: float | None = None,
-    qty: int | None = None,
+    qty: float | None = None,
     notes: str | None = None,
 ) -> bool:
     """부분 업데이트. 존재했으면 True, 없었으면 False.
@@ -205,10 +205,10 @@ def update_holding(
         sets.append("avg_price = ?")
         vals.append(float(avg_price))
     if qty is not None:
-        if int(qty) < 0:
+        if float(qty) < 0:
             raise ValueError(f"qty must be non-negative, got {qty!r}")
         sets.append("qty = ?")
-        vals.append(int(qty))
+        vals.append(float(qty))
     if notes is not None:
         sets.append("notes = ?")
         vals.append(notes)
@@ -239,7 +239,7 @@ def list_holdings(username: str) -> list[dict]:
             (username,),
         ).fetchall()
     return [
-        {"symbol": r[0], "avg_price": float(r[1]), "qty": int(r[2]),
+        {"symbol": r[0], "avg_price": float(r[1]), "qty": float(r[2]),
          "added_at": int(r[3]), "notes": r[4]}
         for r in rows
     ]
@@ -276,7 +276,7 @@ def list_holdings_with_pnl(username: str) -> list[dict]:
     result = []
     for r in rows:
         avg_price = float(r[1])
-        qty = int(r[2])
+        qty = float(r[2])
         last_close = float(r[6]) if r[6] is not None else None
         pnl_pct = None
         pnl_abs = None
@@ -344,7 +344,7 @@ def record_buy(
                 (username, symbol),
             ).fetchone()
             if existing:
-                old_avg, old_qty = float(existing[0]), int(existing[1])
+                old_avg, old_qty = float(existing[0]), float(existing[1])
                 new_qty = old_qty + qty
                 new_avg = (old_avg * old_qty + price * qty) / new_qty
             else:
@@ -401,7 +401,7 @@ def record_sell(
             ).fetchone()
             if not existing:
                 raise ValueError(f"no holding for {username}/{symbol}")
-            old_avg, old_qty = float(existing[0]), int(existing[1])
+            old_avg, old_qty = float(existing[0]), float(existing[1])
             if qty > old_qty:
                 raise ValueError(
                     f"sell qty {qty} exceeds holding {old_qty}")
@@ -440,7 +440,7 @@ def record_sell(
 def record_adjust(
     username: str, symbol: str, *,
     avg_price: float | None = None,
-    qty: int | None = None,
+    qty: float | None = None,
     notes: str | None = None,
 ) -> bool:
     """수동 조정 — update_holding 과 동일하지만 transactions 에 ADJUST 기록.
@@ -458,11 +458,11 @@ def record_adjust(
         vals.append(float(avg_price))
         tx_records.append(("avg_price", float(avg_price)))
     if qty is not None:
-        if int(qty) < 0:
+        if float(qty) < 0:
             raise ValueError(f"qty must be non-negative, got {qty!r}")
         sets.append("qty = ?")
-        vals.append(int(qty))
-        tx_records.append(("qty", float(int(qty))))
+        vals.append(float(qty))
+        tx_records.append(("qty", float(qty)))
     if notes is not None:
         sets.append("notes = ?")
         vals.append(notes)
